% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{age}
\alias{age}
\title{Age}
\usage{
age(from, to)
}
\arguments{
\item{from}{Date (or character vector that can be parsed to Date) to start calculating age, typically date of birth.}

\item{to}{Date on which to determine the age.}
}
\value{
The age in years, integer.
}
\description{
Calculate a subjects age in years.
}
\examples{
age("1990-11-23", Sys.Date())
}
\keyword{format}
\keyword{numeric}
