% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sdtINR.R
\name{calc_sdtINR}
\alias{calc_sdtINR}
\title{Standard deviation of transformed INR's}
\usage{
calc_sdtINR(INR_meas)
}
\arguments{
\item{INR_meas}{Data.frame with (at least) the following columns:
INR_date (a date object with the date of the INR measurement),  
INR (the international normalised ratio).
Optionally it contains a column tt, with the time between measurements in days.}
}
\description{
Calculate sdtINR according to Lind.
}
