% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_ci.R
\name{est_ci}
\alias{est_ci}
\title{Combine an estimate and a confidence interval}
\usage{
est_ci(...)
}
\arguments{
\item{x}{Either an object from lm or glm, or a broom::tidy()-object of those types.}

\item{digits}{Decimals to use in estimate, defaults to 2}

\item{dash}{Separator between the min and max of the confidence interval. Sent to dash()}

\item{CI_text}{Text to preceed the confidence interval, e.g. "95pct CI ".}
}
\value{
Named character vector.
}
\description{
Summarise a glm, lm or a broom::tidy() object of aforementioned types to
an estimate with 95 percent confidence interval for every term in the object.
}
\examples{
x <- lm(log(mpg) ~ log(disp), data = mtcars)
y <- glm( mpg < 20 ~ log(disp), binomial(), data = mtcars)
z <- broom::tidy(y, exponentiate = TRUE, conf.int = TRUE)
est_ci(x)
est_ci(y)
est_ci(z)
}
