% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_numer.R
\name{format_number}
\alias{format_number}
\title{Number formatting}
\usage{
format_number(x, digits = 2)
}
\arguments{
\item{x}{Numeric vector that needs to be formatted.}

\item{digits}{Number of decimals required. Can be a vector the same length as x, or length 1.}
}
\value{
A character vector.
}
\description{
Format a number to contain exactly a number of decimals.
}
\examples{
format_number(pi, 2)
}
\keyword{format}
\keyword{numeric}
