% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_pval.R
\name{format_pval}
\alias{format_pval}
\title{Format a P value}
\usage{
format_pval(x, digits = 2, small = 10^(-1 * (digits + 1)))
}
\arguments{
\item{x}{Numerical vector}

\item{digits}{Digits to show}

\item{small}{Threshold to show "smaller than X"}
}
\value{
Character vector
}
\description{
Formats a p value to a number of digits, or <X
}
\examples{
format_pval(0.0001)
}
