% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_pct.R
\name{n_pct}
\alias{n_pct}
\title{Number and percentages}
\usage{
n_pct(x, digits = 0)
}
\arguments{
\item{x}{Vector with logicals}

\item{digits}{Number of decimals in percentage, defaults to 0}
}
\value{
Number of TRUEs, and percentage of occurrences
}
\description{
Number and percentages
}
\examples{
x <- c(TRUE, TRUE, FALSE)
n_pct(x)
}
