% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/or_to_rr.R
\name{OR_to_RR}
\alias{OR_to_RR}
\title{OR_to_RR}
\usage{
OR_to_RR(OR, p0)
}
\arguments{
\item{OR}{Odds ratio to convert, as a decimal number.}

\item{p0}{Baseline risk, between 0 and 1}
}
\value{
Relative risk, as a decimal number
}
\description{
Convert odds ratio to a relative risk.
}
\examples{
OR_to_RR(10, 0.1)
}
\keyword{format}
\keyword{numeric}
