% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_norm.R
\name{transform_norm}
\alias{transform_norm}
\title{Transform to normal}
\usage{
transform_norm(x)
}
\arguments{
\item{x}{Vector of inputs}
}
\value{
X transformed to a normal distribution.
}
\description{
Force a normal distribution on data.
Uses an empirical cumulative density function for the input data,
and places it on a normal distribution according to the p value.
The mean and sd of the standard distribution are based on the original data.
}
\examples{
transform_norm(runif(100))
}
