% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_sd.R
\name{mean_sd}
\alias{mean_sd}
\title{Summarise a numerical vector to mean (SD).}
\usage{
mean_sd(...)
}
\arguments{
\item{x}{Numerical vector}

\item{digits}{Number of decimals to use}
}
\value{
Character vector
}
\description{
Summarise a numerical vector to a mean and a SD,
or a median and an interquartile range (p25 and p75).
}
\examples{
mean_sd(1:10)
median_iqr(1:10)
}
